+++ src/ADN.Helpers/src/Data/DateTimeHelper.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ADN.Helpers.Data
+{
+    public class DateTimeHelper
+    {
+        public static long DateTimeToUnixTimeStamp(DateTime date)
+        {
+            long unixTimestamp = date.Ticks - new DateTime(1970, 1, 1).Ticks;
+            unixTimestamp /= TimeSpan.TicksPerSecond;
+
+            return unixTimestamp;
+        }
+
+        public static DateTime UnixTimeStampToDateTime(long unixTimestamp)
+        {
+            DateTime date = new DateTime(1970, 1, 1);
+            long unixTimeStampInTicks = unixTimestamp * TimeSpan.TicksPerSecond;
+
+            return new DateTime(date.Ticks + unixTimeStampInTicks);
+        }
+    }
+}