+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Helpers classes: ArrayHelper, ByteHelper.
+- Helpers classes: ArrayHelper, ByteHelper and DateTimeHelper.
 
 ## [0.0.1] - 2018-07-30
 ### Added